requirements.txt
django==2.2.28
djangorestframework==3.14.0
celery==5.3.6
django-celery-beat==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
dj-database-url==2.1.0
requests==2.31.0
